+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-### 0.0.0 - 2016-03-23
+### 0.1.0 - 2017-10-17
 
-#### Added/Changed etc.
+#### Added
 
-* Details
+* `*` support. 
+  * `relation.*` returns false even `relation.*` is whitelisted, but some fields of relation are blacklisted.
+  * `relation.*` returns true when there is only blacklist and relation is not black listed.
+